refactor(storage): move UrlStorageAsPath class in a new file
